LayoutTests/http/tests/security/contentSecurityPolicy/srcdoc-doesnt-bypass-script-src-expected.txt

CONSOLE MESSAGE: Refused to execute inline script because of Content-Security-Policy.

This test passes if it doesn't alert fail. 
